requirements.txt
gradio
python-dotenv
huggingface_hub
Pillow